+++ projects/ng-aquila/src/datefield/datemask/datemask.component.ts
@@ -2,7 +2,6 @@ import { NxFormfieldComponent, NxFormfieldControl } from '@allianz/ng-aquila/for
 import { NxAbstractControl } from '@allianz/ng-aquila/shared';
 import { ErrorStateMatcher, IdGenerationService } from '@allianz/ng-aquila/utils';
 import { FocusMonitor } from '@angular/cdk/a11y';
-
 import {
   AfterContentInit,
   AfterViewInit,
@@ -517,9 +516,16 @@ export class NxDatemaskComponent<D>
     this.moveFocus(this._inputs()[0]);
   }
 
+  private _resetInputValue() {
+    this._dayValue.set(null);
+    this._monthValue.set(null);
+    this._yearValue.set(null);
+  }
+
   writeValue(obj: any): void {
-    const isDate = this._dateAdapter.isDateInstance(obj);
-    if (!isDate) {
+    const isValidDate = this._dateAdapter.isDateInstance(obj) && this._dateAdapter.isValid(obj);
+    if (!isValidDate) {
+      this._resetInputValue();
       return;
     }
 
@@ -625,18 +631,14 @@ export class NxDatemaskComponent<D>
 
     if (event.key === 'Delete' || event.key === 'Backspace') {
       event.preventDefault();
-      this._dayValue.set(null);
-      this._monthValue.set(null);
-      this._yearValue.set(null);
+      this._resetInputValue();
       this.moveFocus(this._inputs()[0], 'start');
     }
 
     if (event.key.length === 1 && event.key.match(/\d/g)) {
       const firstInputValueSignal = this.getValueSignalOfFirstInput();
       event.preventDefault();
-      this._dayValue.set(null);
-      this._monthValue.set(null);
-      this._yearValue.set(null);
+      this._resetInputValue();
       this._inputs()[0].elementRef.nativeElement.value = event.key;
       firstInputValueSignal.set(event.key);
       this.moveFocus(this._inputs()[0], 'none');